move undeltifiedParser and deltifiedParser and entranceFileParser into Entrance class and move Entrance class to the ObjectManager
